update time - STYLIX BROKE and TO DO
after update stylix started to complain.
Stylix is installed in flake as general nix module not home manager.
I use home manager to customize my apps like bat, btop or kitty.
I put together home manager options and nixos options which doesnt work,
so now they are separate - all apps that are excluded from sytlix
styling are put in in stylix.nix and rest of each app config are in
given home manager module like bat.nix

TO DO: move stylix to home manager where it should be from the start and
you could have both options in home manager so it will be clearer and
easier to work with